made tags fragment abstract to support overriding layout.
