+++ WordPress/src/main/java/org/wordpress/android/modules/AppComponent.java
@@ -98,6 +98,7 @@ import org.wordpress.android.ui.posts.PrepublishingActionsFragment;
 import org.wordpress.android.ui.posts.PrepublishingBottomSheetFragment;
 import org.wordpress.android.ui.posts.PublishNotificationReceiver;
 import org.wordpress.android.ui.posts.SelectCategoriesActivity;
+import org.wordpress.android.ui.posts.TagsFragment;
 import org.wordpress.android.ui.posts.adapters.AuthorSelectionAdapter;
 import org.wordpress.android.ui.prefs.AccountSettingsFragment;
 import org.wordpress.android.ui.prefs.AppSettingsFragment;
@@ -495,7 +496,7 @@ public interface AppComponent extends AndroidInjector<WordPress> {
 
     void inject(PrepublishingActionsAdapter object);
 
-    void inject(PostSettingsTagsFragment object);
+    void inject(TagsFragment object);
 
     // Allows us to inject the application without having to instantiate any modules, and provides the Application
     // in the app graph
